requirements.txt
aiosqlite==0.17.0
anyio==3.3.4
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
iso8601==0.1.16
pydantic==1.8.2
pypika-tortoise==0.1.1
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
